requirements.txt
beautifulsoup4==4.10.0
gensim==4.1.2
h5py==3.6.0
jieba==0.42.1
keras==2.7.0
matplotlib==3.7.2
numpy==1.21.1
pandas==2.0.3
pytz==2022.7
PyYAML==6.0.1
scikit_learn==1.2.0
tensorflow_addons==0.22.0
tensorflow_gpu==2.7.0
torch==1.13.1
